eventAggregator: split away the eventReducer
